IdeaInspiration/Sources/Events/SportsHighlights/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "prismq-idea-sources-events-sportshighlights"
version = "1.0.0"
description = "PrismQ module for gathering event inspirations from sports events and highlights"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "PrismQ", email = "dev@prismq.com"}
]
keywords = ["prismq", "ai", "content-generation", "events", "sports", "idea-sources"]

dependencies = [
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
sports-highlights-source = "src.cli:main"

[tool.setuptools]
packages = ["src", "src.core", "src.plugins"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
